Changed short to int to avoid overflow when checking maxval

+++ headers/ppm.h
@@ -5,7 +5,7 @@
 struct image
 {
   unsigned int height, width;
-  short maxval;
+  int maxval;
   matrix img;
 };
